resource service fixed

+++ KnowledgeBase.Logic/Services/ResourceService.cs
@@ -3,57 +3,65 @@ using KnowledgeBase.Data.Models;
 using KnowledgeBase.Data.Repositories.Interfaces;
 using KnowledgeBase.Logic.Dto;
 using KnowledgeBase.Logic.Services.Interfaces;
-using Microsoft.EntityFrameworkCore.Storage.ValueConversion.Internal;
+using KnowledgeBase.Shared;
 
-namespace KnowledgeBase.Logic.Services
+namespace KnowledgeBase.Logic.Services;
+
+public class ResourceService : IResourceService
 {
-    public class ResourceService : IResourceService
+    private readonly IResourceRepository _resourceRepository;
+    private readonly IMapper _mapper;
+
+    public ResourceService(IResourceRepository resourceService, IMapper mapper)
     {
-        private readonly IResourceRepository _resourceRepository;
-		private readonly IMapper _mapper;
+        _resourceRepository = resourceService;
+        _mapper = mapper;
+    }
 
-		public ResourceService(IResourceRepository resourceService, IMapper mapper)
-        {
-            _resourceRepository = resourceService;
-			_mapper = mapper;
-		}
+    public ResourceDto? Get(Guid id)
+    {
+        var resource = _resourceRepository.Get(id);
+        return _mapper.Map<ResourceDto>(resource);
+    }
 
-		public Resource Get(Guid id)
+    public void Add(ResourceDto resourcedto)
+    {
+        Resource resource = _mapper.Map<Resource>(resourcedto);
+        _resourceRepository.Add(resource);
+    }
+
+    public void Remove(ResourceDto resourcedto)
+    {
+        Resource resource = _mapper.Map<Resource>(resourcedto);
+        _resourceRepository.Remove(resource);
+    }
+
+    public void Delete(ResourceDto resourceDto)
+    {
+        var id = resourceDto.Id.ToGuid();
+        if (id == Guid.Empty)
         {
-            return _resourceRepository.Get(id);
+            return;
         }
 
-        public void Add(ResourceDto resourcedto)
+        var resource = _resourceRepository.Get(id);
+        if (resource == null) // Project doesnt exist
         {
-			Resource resource = _mapper.Map<Resource>(resourcedto);
-			_resourceRepository.Add(resource);
+            return;
         }
 
-        public void Remove(ResourceDto resourcedto)
-        {
-			Resource resource = _mapper.Map<Resource>(resourcedto);
-			_resourceRepository.Remove(resource);
-        }
+        _resourceRepository.Delete(resource);
+    }
 
-        public void Delete(ResourceDto resourcedto)
-        {
-			Resource resource = Get(_mapper.Map<Resource>(resourcedto).Id);
-			_resourceRepository.Delete(resource);
-        }
+    public void Update(ResourceDto resourcedto)
+    {
+        Resource resource = _mapper.Map<Resource>(resourcedto);
+        _resourceRepository.Update(resource);
+    }
 
-        public void Update(ResourceDto resourcedto)
-        {
-			Resource resource = _mapper.Map<Resource>(resourcedto);
-			_resourceRepository.Update(resource);
-        }
-
-        public IEnumerable<Resource> GetAll()
-        {
-            IEnumerable<Resource> resourceList = _resourceRepository.GetAll();
-            return resourceList.ToList();
-        }
+    public IEnumerable<ResourceDto> GetAll()
+    {
+        IEnumerable<Resource> resourceList = _resourceRepository.GetAll();
+        return resourceList.Select(r => _mapper.Map<ResourceDto>(r)).ToList();
     }
 }
-
-
-
